move date specific util fns to date util file
